requirements.txt
asgiref==3.6.0
contextlib2==21.6.0
Django==4.1.7
django-ninja==0.21.0
django-ninja-extra==0.18.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
injector==0.20.1
pydantic==1.10.6
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
